requirements.txt
grpcio==1.14.1
grpcio-tools==1.14.1
gcloud==0.18.3
google-api-python-client==1.7.3
metasdk==0.0.3